Moving CubeBackgroundModel class from examples to background.models.
